refactor : 위도, 경도 데이터 저장을 위한 타입 변경 & 테스트 객체 생성을 위해 dto에 Builder 어노테이션 추가

+++ src/main/java/pp/coinwash/user/domain/dto/CustomerResponseDto.java
@@ -10,8 +10,8 @@ public record CustomerResponseDto(
 	String phone,
 	int points,
 	String address,
-	Long latitude,
-	Long longitude) {
+	Double latitude,
+	Double longitude) {
 
 	public static CustomerResponseDto from(Customer customer) {
 		return CustomerResponseDto.builder()
